few improvements
line 18-21 on Windows, this pull request introduces new functionality. It checks if the user is inactive (no mouse/keyboard usage) and if so, it goes to sleep. I use it, as quite often I fall asleep or leave, and the less prolific's API is used, the less it costs, the less likely they will ban us/or disable it

- line 120 it works as it is. It goes straight to the study, therefore there is no # TODO

- line 145-170 script pausing functionality implementation. Most likely this is impossible to get done under PyAutoGui, as it's too dumb, I checked. Shame, as PyAutoGui would make it cross platform (Linux, Mac)

lines 162-167 unnecessary

lines 191-204 implementation of script sleeping on the Windows platform